riscv: cheri: Make various library routines cheri-ready
- Implement user_ptr.h for riscv-cheri
- Adapt vsprintf implementation to riscv-cheri and to a
  CHERI enabled kernel.
- Make various library routines ready for a CHERI enabled kernel.
  In particular this affects data structures that store additional
  bits in pointers.
- Handle all CHERI and non-CHERI cases in alignment macros for
  pointers ({USER_,}PTR_ALIGN{,_DOWN}).

+++ lib/scatterlist.c
@@ -1223,7 +1223,7 @@ static ssize_t extract_kvec_to_sg(struct iov_iter *iter,
 
 	for (i = 0; i < iter->nr_segs; i++) {
 		struct page *page;
-		unsigned long kaddr;
+		uintptr_t kaddr;
 		size_t off, len, seg;
 
 		len = kv[i].iov_len;
